Try to avoid missing new petitions
Apparently the petitions do not always appear on the site in strict order of
"created" time -- which is not terribly surprising.

This means we sometimes miss new petitions, especially during busy periods.

Try to make this happen less often, by requiring a certain number of consecutive
petitions that we have already seen before we assume we’ve already seen all the
following.

This is probably not a great solution. Or in any case the MAX_CONSECUTIVE parameter
may need to be increased.